Removed time.sleep(2) after LED came on; renamed
This line of code made sense when the original program was turning on an LED in response to movement; so far it has just been causing a little hink in the process whereby the voice takes a second (well, two) to kick in. HOPEFULLY this didn't break anything, though I can't see what it would have done. In practice the LED was staying on throughout the audio file so it became vestigial.

Also renamed file.